code refactoring, docs updates.
