+++ README.md
@@ -33,13 +33,15 @@ var text = "var t; // comments";
 decomment(text); //=> var t;
 ```
 
+NOTE: Specifically for CSS, call `decomment.css(text, [options])` instead.
+
 ## Features
 
 * Removes both single and multi-line comments from JSON, JavaScript and CSS
 * Automatically recognizes HTML and removes all `<!-- comments -->` from it
 * Does not change layout / formatting of the original document
 * Removes lines that have only comments on them
-* Compatible with JSON5 and ECMAScript 6
+* Compatible with CSS3, JSON5 and ECMAScript 6
 
 The library does not support mixed content - HTML with JavaScript or CSS in it.
 Once the input code is recognized as HTML, only the HTML comments will be removed from it.
@@ -48,12 +50,11 @@ Once the input code is recognized as HTML, only the HTML comments will be remove
 
 In terms of the performance, this library is as fast as it gets, in part because it makes no use of regular expressions.
 
-For example, it churns through [AngularJS 1.5 Core](https://code.angularjs.org/1.5.0-rc.0/angular.js) (1.1MB ~ 30,000 lines of JavaScript) in under 40ms. 
+For example, it churns through [AngularJS 1.5 Core](https://code.angularjs.org/1.5.0-rc.0/angular.js) (1.1MB ~ 30,000 lines of JavaScript) in under 200ms. 
 
 ## API
 
 #### decomment(text, [options]) ⇒ String
-
 
 ##### options.trim ⇒ Boolean
 * `false (default)` - do not trim comments
@@ -81,7 +82,11 @@ decomment(text, {safe: true}); //=> /*! special */ js code
 
 This option has no effect when processing HTML.
 
+#### decomment.css(text, [options]) ⇒ String
+
+The same as **decomment**, but specific to CSS.
+
 ## License
 
-Copyright © 2015 [Vitaly Tomilov](https://github.com/vitaly-t);
+Copyright © 2016 [Vitaly Tomilov](https://github.com/vitaly-t);
 Released under the MIT license.
